Moves ConvexHull out from MultiShapeConf
