+++ Library/CMakeLists.txt
@@ -78,6 +78,7 @@ set(PLAYRHO_D2_HDRS
 	"${PLAYRHO_INCLUDE_PREFIX}d2/CodeDumper.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/ContactImpulsesList.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/ContactSolver.hpp"
+	"${PLAYRHO_INCLUDE_PREFIX}d2/ConvexHull.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/DiskShapeConf.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/Distance.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/DistanceConf.hpp"
@@ -144,6 +145,7 @@ set(PLAYRHO_D2_SRCS
 	"${PLAYRHO_SOURCE_PREFIX}d2/CodeDumper.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/ContactImpulsesList.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/ContactSolver.cpp"
+	"${PLAYRHO_SOURCE_PREFIX}d2/ConvexHull.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/DiskShapeConf.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/Distance.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/DistanceConf.cpp"
